+++ Tests/QuillCodeParityTests/ParityDesktopBrowserAdapterGateTests.swift
@@ -54,6 +54,9 @@ final class ParityDesktopBrowserAdapterGateTests: QuillCodeParityTestCase {
         let browserCoordinatorText = try Self.desktopSourceText(named: "QuillCodeDesktopBrowserCoordinator.swift")
         let appText = try Self.desktopSourceText(named: "QuillCodeDesktopApp.swift")
         let presenterText = try Self.desktopSourceText(named: "DesktopBrowserSessionPresenter.swift")
+        let presenterContractsText = try Self.desktopSourceText(named: "DesktopBrowserSessionContracts.swift")
+        let sessionJavaScriptText = try Self.desktopSourceText(named: "DesktopBrowserSessionJavaScript.swift")
+        let sessionWindowText = try Self.desktopSourceText(named: "DesktopBrowserSessionWindowController.swift")
         let extractorText = try Self.desktopSourceText(named: "DesktopBrowserLiveDOMSnapshotExtractor.swift")
         let browserPaneText = try Self.appSourceText(named: "QuillCodeBrowserPaneView.swift")
         let browserControlsText = try Self.appSourceText(named: "QuillCodeBrowserPaneControls.swift")
@@ -61,10 +64,10 @@ final class ParityDesktopBrowserAdapterGateTests: QuillCodeParityTestCase {
         let viewCommandPlannerText = try Self.appSourceText(named: "QuillCodeWorkspaceViewCommandPlanner.swift")
 
         Self.assertSource(desktopText, contains: "DesktopBrowserSessionPresenter")
-        Self.assertSource(presenterText, contains: "protocol DesktopBrowserSessionPresenting")
-        Self.assertSource(presenterText, contains: "WKWebView")
-        Self.assertSource(presenterText, contains: "configuration.websiteDataStore = .default()")
-        Self.assertSource(presenterText, contains: "loadFileURL")
+        Self.assertSource(presenterContractsText, contains: "protocol DesktopBrowserSessionPresenting")
+        Self.assertSource(sessionWindowText, contains: "WKWebView")
+        Self.assertSource(sessionWindowText, contains: "configuration.websiteDataStore = .default()")
+        Self.assertSource(sessionWindowText, contains: "loadFileURL")
         Self.assertSource(presenterText, contains: "private var session: DesktopBrowserSessionWindowController?")
         Self.assertSource(presenterText, contains: "presentSession(_ snapshot: BrowserSessionSyncSnapshot)")
         Self.assertSource(presenterText, contains: "syncSession(_ snapshot: BrowserSessionSyncSnapshot)")
@@ -82,21 +85,21 @@ final class ParityDesktopBrowserAdapterGateTests: QuillCodeParityTestCase {
         Self.assertSource(presenterText, contains: "session.clickInSelectedTab(selector: selector)")
         Self.assertSource(presenterText, contains: "session.typeInSelectedTab(selector: selector, text: text, submit: submit)")
         Self.assertSource(presenterText, contains: "session?.reloadSelectedTab()")
-        Self.assertSource(presenterText, contains: "tab.webView.goBack()")
-        Self.assertSource(presenterText, contains: "tab.webView.goForward()")
-        Self.assertSource(presenterText, contains: "tab.webView.evaluateJavaScript(trimmedSource)")
-        Self.assertSource(presenterText, contains: "document.querySelector(selector)")
-        Self.assertSource(presenterText, contains: "element.click()")
-        Self.assertSource(presenterText, contains: "element.dispatchEvent(new InputEvent")
-        Self.assertSource(presenterText, contains: "JSONEncoder().encode(value)")
-        Self.assertSource(presenterText, contains: "DesktopBrowserLiveDOMSnapshotExtractor.snapshot")
-        Self.assertSource(presenterText, contains: "tab.webView.reload()")
-        Self.assertSource(presenterText, contains: "session.sync(snapshot)")
-        Self.assertSource(presenterText, contains: "emitRenderedSessionUpdate")
-        Self.assertSource(presenterText, contains: "liveDOMSnapshot: liveDOMSnapshots[id]")
+        Self.assertSource(sessionWindowText, contains: "tab.webView.goBack()")
+        Self.assertSource(sessionWindowText, contains: "tab.webView.goForward()")
+        Self.assertSource(sessionWindowText, contains: "tab.webView.evaluateJavaScript(trimmedSource)")
+        Self.assertSource(sessionJavaScriptText, contains: "document.querySelector(selector)")
+        Self.assertSource(sessionJavaScriptText, contains: "element.click()")
+        Self.assertSource(sessionJavaScriptText, contains: "element.dispatchEvent(new InputEvent")
+        Self.assertSource(sessionJavaScriptText, contains: "JSONEncoder().encode(value)")
+        Self.assertSource(sessionWindowText, contains: "DesktopBrowserLiveDOMSnapshotExtractor.snapshot")
+        Self.assertSource(sessionWindowText, contains: "tab.webView.reload()")
+        Self.assertSource(sessionWindowText, contains: "sync(snapshot)")
+        Self.assertSource(sessionWindowText, contains: "emitRenderedSessionUpdate")
+        Self.assertSource(sessionWindowText, contains: "liveDOMSnapshot: liveDOMSnapshots[id]")
         Self.assertSource(extractorText, contains: "DesktopBrowserLiveDOMSnapshotExtractor")
-        Self.assertSource(presenterText, contains: "NSTabView")
-        Self.assertSource(presenterText, contains: "func present()")
+        Self.assertSource(sessionWindowText, contains: "NSTabView")
+        Self.assertSource(sessionWindowText, contains: "func present()")
         Self.assertSource(presenterText, excludes: "sessions: [ObjectIdentifier")
         Self.assertSource(presenterText, excludes: "func openSession(url: URL)")
         Self.assertSource(controllerText, contains: "browserSessionPresenter")
